ensure we have small video in order to take screenshots
