+++ controllers/url-helpers.js
@@ -1,0 +1,6 @@
+function youtubeUrlFromId(ytid) {
+  return `https://www.youtube.com/watch?v=${ytid}`;
+}
+module.exports = {
+  youtubeUrlFromId,
+};
